creditcoin: move ocw helpers into ocw mod
